Lisää local-shibbo profiili jolla voi testata shibboleth asioita paikallisesti
