+++ src/test/local-shibbo/default.properties
@@ -1,0 +1,25 @@
+saml.keystoreLocation=src/test/local-shibbo/keystore.jks
+saml.keystorePassword=salasana
+saml.idp.metadataUrl=https://login-test.it.helsinki.fi/metadata/sign-hy-test-metadata.xml
+saml.sp.scheme=http
+saml.student.domain=1
+saml.student.metadataLocation=src/test/local-shibbo/local-student-metadata.xml
+saml.student.alias=student
+
+saml.teacher.domain=local.teacher.helsinki.fi
+saml.teacher.metadataLocation=src/test/local-shibbo/local-teacher-metadata.xml
+saml.teacher.alias=teacher
+ssoLogoutUri=https://login-test.it.helsinki.fi/idp/profile/Logout
+environment=qa
+logging.level.org.springframework.security=DEBUG
+
+hostUrl=https://local.student.helsinki.fi:3000
+loginUrlTeacher=/saml/login/alias/teacher
+loginUrlStudent=/saml/login/alias/student
+logoutUrl=saml/logout
+teacherAppUrl=https://local.teacher.helsinki.fi:3000
+studentAppUrl=https://local.student.helsinki.fi:3000
+student=https://local.student.helsinki.fi:3002/profile
+teacher=https://local.teacher.helsinki.fi:3002/profile
+
+whitelistedIps=127.0.0.1
